reverted simple structures and cleaned a bit dgeoflow
