+++ applications/GeoMechanicsApplication/custom_workflows/adaptive_time_incrementor.cpp
@@ -12,7 +12,7 @@
 //
 
 #include "adaptive_time_incrementor.h"
-#include "time_step_end_state.h"
+#include "time_step_end_state.hpp"
 
 #include "includes/exception.h"
