+++ applications/GeoMechanicsApplication/custom_workflows/strategy_wrapper.h
@@ -15,7 +15,7 @@
 
 #include <cstddef>
 
-#include "time_step_end_state.h"
+#include "time_step_end_state.hpp"
 
 namespace Kratos
 {
