+++ applications/GeoMechanicsApplication/custom_workflows/time_incrementor.h
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include "time_step_end_state.h"
+#include "time_step_end_state.hpp"
 
 #include <optional>
