+++ applications/GeoMechanicsApplication/custom_workflows/time_loop_executor_interface.h
@@ -18,7 +18,7 @@
 #include <memory>
 #include <vector>
 
-#include "time_step_end_state.h"
+#include "time_step_end_state.hpp"
 
 namespace Kratos
 {
