+++ applications/GeoMechanicsApplication/custom_workflows/time_step_executor.h
@@ -15,7 +15,7 @@
 
 #include "processes/process.h"
 #include "strategy_wrapper.h"
-#include "time_step_end_state.h"
+#include "time_step_end_state.hpp"
 
 #include <functional>
 #include <memory>
