+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_workflows/test_prescribed_time_incrementor.cpp
@@ -12,7 +12,7 @@
 //
 
 #include "custom_workflows/prescribed_time_incrementor.h"
-#include "custom_workflows/time_step_end_state.h"
+#include "custom_workflows/time_step_end_state.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 
 using namespace Kratos;
